Centralizing Koza mock classes for sharing
